GitHub gitignore bypass
Since GitHub automatically adds "/bin" at the end of the gitignore file
when we change things a little, we will rename "bin" into "files", and
add all the other files that are supposed to belong to the future final
JAR package.
Do not forget to change the Run Configuration with
"-splash:files/splash.png".